docker/encoders/gcn/requirements.txt
fastapi
uvicorn
torch
torch-geometric
rdkit-pypi
numpy